Refactor: move optimization logic to separate modules
