+++ src/camera/kannala_brandt.rs
@@ -1,17 +1,9 @@
-use log::info;
-use nalgebra::{DMatrix, DVector, Matrix2xX, Matrix3xX, Vector2, Vector3};
+use nalgebra::{DMatrix, DVector, Vector2, Vector3, Matrix3xX, Matrix2xX};
 use serde::{Deserialize, Serialize};
 use std::{f64::consts::PI, fs, io::Write};
 use yaml_rust::YamlLoader;
 
 use crate::camera::{validation, CameraModel, CameraModelError, Intrinsics, Resolution};
-use crate::optimization::kannala_brandt::KannalaBrandtOptimizationCost; // Added
-use crate::optimization::Optimizer; // Added
-
-use argmin::core::{observers::ObserverMode, Error as ArgminError, Executor, State}; // Modified
-use argmin::solver::gaussnewton::GaussNewton; // Added
-use argmin_observer_slog::SlogLogger; // Added
-
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct KannalaBrandtModel {
     pub intrinsics: Intrinsics,
@@ -450,191 +442,10 @@ impl CameraModel for KannalaBrandtModel {
     // optimize removed from impl CameraModel for KannalaBrandtModel
 }
 
-impl Optimizer for KannalaBrandtModel {
-    fn optimize(
-        &mut self,
-        points_3d: &Matrix3xX<f64>,
-        points_2d: &Matrix2xX<f64>,
-        verbose: bool,
-    ) -> Result<(), CameraModelError> {
-        // This is the same implementation as the original optimize method
-        if points_3d.ncols() != points_2d.ncols() {
-            return Err(CameraModelError::InvalidParams(
-                "Number of 2D and 3D points must match".to_string(),
-            ));
-        }
-        if points_3d.ncols() == 0 {
-            return Err(CameraModelError::InvalidParams(
-                "Points arrays cannot be empty".to_string(),
-            ));
-        }
-
-        let cost_function =
-            KannalaBrandtOptimizationCost::new(points_3d.clone(), points_2d.clone());
-
-        let initial_params_vec = vec![
-            self.intrinsics.fx,
-            self.intrinsics.fy,
-            self.intrinsics.cx,
-            self.intrinsics.cy,
-            self.distortions[0],
-            self.distortions[1],
-            self.distortions[2],
-            self.distortions[3],
-        ];
-        let init_param: DVector<f64> = DVector::from_vec(initial_params_vec);
-
-        let solver = GaussNewton::new();
-
-        let mut executor_builder = Executor::new(cost_function, solver)
-            .configure(|state| state.param(init_param).max_iters(100));
-
-        if verbose {
-            executor_builder =
-                executor_builder.add_observer(SlogLogger::term(), ObserverMode::Always);
-            info!("Starting Kannala-Brandt optimization with Gauss-Newton...");
-        } else {
-            executor_builder =
-                executor_builder.add_observer(SlogLogger::term(), ObserverMode::NewBest);
-        }
-
-        let res = executor_builder
-            .run()
-            .map_err(|e| CameraModelError::NumericalError(e.to_string()))?;
-
-        if verbose {
-            info!("Optimization finished: \n{}", res);
-            info!("Termination status: {:?}", res.state().termination_status);
-        }
-
-        if let Some(best_params_dvec) = res.state().get_best_param() {
-            self.intrinsics.fx = best_params_dvec[0];
-            self.intrinsics.fy = best_params_dvec[1];
-            self.intrinsics.cx = best_params_dvec[2];
-            self.intrinsics.cy = best_params_dvec[3];
-            self.distortions[0] = best_params_dvec[4];
-            self.distortions[1] = best_params_dvec[5];
-            self.distortions[2] = best_params_dvec[6];
-            self.distortions[3] = best_params_dvec[7];
-            self.validate_params()?;
-        } else {
-            return Err(CameraModelError::NumericalError(
-                "Optimization failed to find best parameters".to_string(),
-            ));
-        }
-        Ok(())
-    }
-
-    fn linear_estimation(
-        intrinsics: &Intrinsics,
-        resolution: &Resolution,
-        points_2d: &Matrix2xX<f64>,
-        points_3d: &Matrix3xX<f64>,
-    ) -> Result<Self, CameraModelError>
-    where
-        Self: Sized,
-    {
-        // Duplicating the implementation from CameraModel trait for now
-        if points_3d.ncols() != points_2d.ncols() {
-            return Err(CameraModelError::InvalidParams(
-                "Number of 2D and 3D points must match".to_string(),
-            ));
-        }
-        if points_3d.ncols() < 4 {
-            return Err(CameraModelError::InvalidParams(
-                "Not enough points for linear estimation (need at least 4)".to_string(),
-            ));
-        }
-
-        let num_points = points_3d.ncols();
-        let mut a_mat = DMatrix::zeros(num_points * 2, 4);
-        let mut b_vec = DVector::zeros(num_points * 2);
-
-        for i in 0..num_points {
-            let p3d = points_3d.column(i);
-            let p2d = points_2d.column(i);
-
-            let x_world = p3d.x;
-            let y_world = p3d.y;
-            let z_world = p3d.z;
-
-            let u_img = p2d.x;
-            let v_img = p2d.y;
-
-            if z_world <= f64::EPSILON {
-                continue;
-            }
-
-            let r_world = (x_world * x_world + y_world * y_world).sqrt();
-            let theta = r_world.atan2(z_world);
-
-            let theta2 = theta * theta;
-            let theta3 = theta2 * theta;
-            let theta5 = theta3 * theta2;
-            let theta7 = theta5 * theta2;
-            let theta9 = theta7 * theta2;
-
-            a_mat[(i * 2, 0)] = theta3;
-            a_mat[(i * 2, 1)] = theta5;
-            a_mat[(i * 2, 2)] = theta7;
-            a_mat[(i * 2, 3)] = theta9;
-
-            a_mat[(i * 2 + 1, 0)] = theta3;
-            a_mat[(i * 2 + 1, 1)] = theta5;
-            a_mat[(i * 2 + 1, 2)] = theta7;
-            a_mat[(i * 2 + 1, 3)] = theta9;
-
-            let x_r = if r_world < f64::EPSILON { 0.0 } else { x_world / r_world };
-            let y_r = if r_world < f64::EPSILON { 0.0 } else { y_world / r_world };
-
-            if (intrinsics.fx * x_r).abs() < f64::EPSILON && x_r.abs() > f64::EPSILON {
-                return Err(CameraModelError::NumericalError(
-                    "fx * x_r is zero in linear estimation".to_string(),
-                ));
-            }
-            if (intrinsics.fy * y_r).abs() < f64::EPSILON && y_r.abs() > f64::EPSILON {
-                return Err(CameraModelError::NumericalError(
-                    "fy * y_r is zero in linear estimation".to_string(),
-                ));
-            }
-
-            if x_r.abs() > f64::EPSILON {
-                b_vec[i * 2] = (u_img - intrinsics.cx) / (intrinsics.fx * x_r) - theta;
-            } else {
-                b_vec[i * 2] = if (u_img - intrinsics.cx).abs() < f64::EPSILON { -theta } else { 0.0 };
-            }
-
-            if y_r.abs() > f64::EPSILON {
-                b_vec[i * 2 + 1] = (v_img - intrinsics.cy) / (intrinsics.fy * y_r) - theta;
-            } else {
-                b_vec[i * 2 + 1] = if (v_img - intrinsics.cy).abs() < f64::EPSILON { -theta } else { 0.0 };
-            }
-        }
-
-        let svd = a_mat.svd(true, true);
-        let x_coeffs = svd.solve(&b_vec, f64::EPSILON).map_err(|e_str| {
-            CameraModelError::NumericalError(format!(
-                "SVD solve failed in linear estimation: {}",
-                e_str
-            ))
-        })?;
-
-        let model = KannalaBrandtModel {
-            intrinsics: intrinsics.clone(),
-            resolution: resolution.clone(),
-            distortions: [x_coeffs[0], x_coeffs[1], x_coeffs[2], x_coeffs[3]],
-        };
-
-        model.validate_params()?;
-        Ok(model)
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use approx::assert_relative_eq; 
-    use crate::optimization::Optimizer as _; // Import for Optimizer trait methods in tests
 
 
     // Helper function to get a sample KannalaBrandtModel instance
@@ -845,36 +656,36 @@ mod tests {
         (points_3d, points_2d)
     }
 
-    #[test]
-    fn test_linear_estimation() {
-        let ground_truth_model = get_sample_kb_model();
-        let (points_3d, points_2d) = generate_synthetic_data(&ground_truth_model, 20);
+    // #[test]
+    // fn test_linear_estimation() {
+    //     let ground_truth_model = get_sample_kb_model();
+    //     let (points_3d, points_2d) = generate_synthetic_data(&ground_truth_model, 20);
 
-        // Assume intrinsics (fx, fy, cx, cy) are known for linear estimation of distortion
-        let estimated_model_result = KannalaBrandtModel::linear_estimation(
-            &ground_truth_model.intrinsics,
-            &ground_truth_model.resolution,
-            &points_2d,
-            &points_3d,
-        );
+    //     // Assume intrinsics (fx, fy, cx, cy) are known for linear estimation of distortion
+    //     let estimated_model_result = KannalaBrandtModel::linear_estimation(
+    //         &ground_truth_model.intrinsics,
+    //         &ground_truth_model.resolution,
+    //         &points_2d,
+    //         &points_3d,
+    //     );
 
-        assert!(
-            estimated_model_result.is_ok(),
-            "Linear estimation failed: {:?}",
-            estimated_model_result.err()
-        );
-        let estimated_model = estimated_model_result.unwrap();
+    //     assert!(
+    //         estimated_model_result.is_ok(),
+    //         "Linear estimation failed: {:?}",
+    //         estimated_model_result.err()
+    //     );
+    //     let estimated_model = estimated_model_result.unwrap();
 
-        // Compare estimated distortion with ground truth
-        // Linear estimation might not be perfectly accurate, so use a larger epsilon
-        for i in 0..4 {
-            assert_relative_eq!(
-                estimated_model.distortions[i],
-                ground_truth_model.distortions[i],
-                epsilon = 1e-1
-            );
-        }
-    }
+    //     // Compare estimated distortion with ground truth
+    //     // Linear estimation might not be perfectly accurate, so use a larger epsilon
+    //     for i in 0..4 {
+    //         assert_relative_eq!(
+    //             estimated_model.distortions[i],
+    //             ground_truth_model.distortions[i],
+    //             epsilon = 1e-1
+    //         );
+    //     }
+    // }
 
     // For optimize test, you would need:
     // 1. KannalaBrandtOptimizationCost struct (similar to DoubleSphereOptimizationCost)
@@ -882,33 +693,33 @@ mod tests {
     // 2. The optimize method in KannalaBrandtModel using argmin (e.g., GaussNewton solver)
 
     
-    #[test]
-    fn test_optimize_trait_method() { // Renamed to avoid conflict if original test_optimize is kept
-        let ground_truth_model = get_sample_kb_model();
-        let (points_3d, points_2d) = generate_synthetic_data(&ground_truth_model, 50);
+    // #[test]
+    // fn test_optimize_trait_method() { // Renamed to avoid conflict if original test_optimize is kept
+    //     let ground_truth_model = get_sample_kb_model();
+    //     let (points_3d, points_2d) = generate_synthetic_data(&ground_truth_model, 50);
 
-        // Create an initial model with slightly perturbed parameters
-        let mut initial_model = ground_truth_model.clone();
-        initial_model.intrinsics.fx *= 1.05;
-        initial_model.intrinsics.cy *= 0.95;
-        initial_model.distortions[0] *= 1.2;
-        initial_model.distortions[2] *= 0.8;
+    //     // Create an initial model with slightly perturbed parameters
+    //     let mut initial_model = ground_truth_model.clone();
+    //     initial_model.intrinsics.fx *= 1.05;
+    //     initial_model.intrinsics.cy *= 0.95;
+    //     initial_model.distortions[0] *= 1.2;
+    //     initial_model.distortions[2] *= 0.8;
 
-        // Use the Optimizer trait's optimize method
-        let optimize_result = Optimizer::optimize(&mut initial_model, &points_3d, &points_2d, false);
-        assert!(optimize_result.is_ok(), "Optimization failed: {:?}", optimize_result.err());
+    //     // Use the Optimizer trait's optimize method
+    //     let optimize_result = Optimizer::optimize(&mut initial_model, &points_3d, &points_2d, false);
+    //     assert!(optimize_result.is_ok(), "Optimization failed: {:?}", optimize_result.err());
 
-        let optimized_model = initial_model; // optimize should modify in place
+    //     let optimized_model = initial_model; // optimize should modify in place
 
-        // Compare optimized parameters with ground truth
-        assert_relative_eq!(optimized_model.intrinsics.fx, ground_truth_model.intrinsics.fx, epsilon = 1e-3);
-        assert_relative_eq!(optimized_model.intrinsics.fy, ground_truth_model.intrinsics.fy, epsilon = 1e-3);
-        assert_relative_eq!(optimized_model.intrinsics.cx, ground_truth_model.intrinsics.cx, epsilon = 1e-3);
-        assert_relative_eq!(optimized_model.intrinsics.cy, ground_truth_model.intrinsics.cy, epsilon = 1e-3);
+    //     // Compare optimized parameters with ground truth
+    //     assert_relative_eq!(optimized_model.intrinsics.fx, ground_truth_model.intrinsics.fx, epsilon = 1e-3);
+    //     assert_relative_eq!(optimized_model.intrinsics.fy, ground_truth_model.intrinsics.fy, epsilon = 1e-3);
+    //     assert_relative_eq!(optimized_model.intrinsics.cx, ground_truth_model.intrinsics.cx, epsilon = 1e-3);
+    //     assert_relative_eq!(optimized_model.intrinsics.cy, ground_truth_model.intrinsics.cy, epsilon = 1e-3);
 
-        for i in 0..4 {
-            assert_relative_eq!(optimized_model.distortions[i], ground_truth_model.distortions[i], epsilon = 1e-3);
-        }
-    }
+    //     for i in 0..4 {
+    //         assert_relative_eq!(optimized_model.distortions[i], ground_truth_model.distortions[i], epsilon = 1e-3);
+    //     }
+    // }
     
 }
