+++ src/optimization/double_sphere.rs
@@ -1,9 +1,14 @@
 // src/optimization/double_sphere.rs
 
-use crate::camera::double_sphere::DoubleSphereModel; // Adjusted path
-use crate::camera::CameraModel; // For project method
-use crate::optimization::OptimizationCost; // Import the new trait
-use argmin::core::{CostFunction, Error, Gradient, Hessian, Jacobian, Operator};
+// use crate::camera::double_sphere::DoubleSphereModel; // Adjusted path
+use crate::camera::{CameraModel, DoubleSphereModel, CameraModelError, Intrinsics, Resolution};
+use crate::optimization::Optimizer;
+use argmin::{
+    core::{CostFunction, Error, Gradient, Hessian, Jacobian, Operator, observers::ObserverMode, Executor, State},
+    solver::gaussnewton::GaussNewton
+};
+// use argmin::solver::gaussnewton::GaussNewton; // Added
+use argmin_observer_slog::SlogLogger; // Added
 use log::info;
 use nalgebra::{DMatrix, DVector, Matrix2xX, Matrix3xX};
 
@@ -14,6 +19,7 @@ use nalgebra::{DMatrix, DVector, Matrix2xX, Matrix3xX};
 /// for use with the argmin optimization library.
 #[derive(Clone)]
 pub struct DoubleSphereOptimizationCost {
+    model: DoubleSphereModel,
     /// 3D points in camera coordinate system (3×N matrix)
     points3d: Matrix3xX<f64>,
     /// Corresponding 2D points in image coordinates (2×N matrix)
@@ -31,17 +37,148 @@ impl DoubleSphereOptimizationCost {
     /// # Panics
     ///
     /// Panics if the number of 3D and 2D points don't match.
-    pub fn new(points3d: Matrix3xX<f64>, points2d: Matrix2xX<f64>) -> Self {
+    pub fn new(model: DoubleSphereModel, points3d: Matrix3xX<f64>, points2d: Matrix2xX<f64>) -> Self {
         assert_eq!(points3d.ncols(), points2d.ncols());
-        DoubleSphereOptimizationCost { points3d, points2d }
+        DoubleSphereOptimizationCost { model, points3d, points2d }
     }
 }
 
-impl OptimizationCost for DoubleSphereOptimizationCost {
-    type Param = DVector<f64>;
-    type Output = DVector<f64>; // For Operator
-    type Jacobian = DMatrix<f64>; // For Jacobian
-    type Hessian = DMatrix<f64>; // For Hessian
+impl Optimizer for DoubleSphereOptimizationCost {
+    fn optimize(
+        &mut self,
+        points_3d: &Matrix3xX<f64>,
+        points_2d: &Matrix2xX<f64>,
+        verbose: bool,
+    ) -> Result<(), CameraModelError> {
+        if points_3d.ncols() != points_2d.ncols() {
+            return Err(CameraModelError::InvalidParams(
+                "Number of 2D and 3D points must match".to_string(),
+            ));
+        }
+
+        if points_3d.ncols() == 0 {
+            return Err(CameraModelError::InvalidParams(
+                "Points arrays cannot be empty".to_string(),
+            ));
+        }
+
+        let cost_function = DoubleSphereOptimizationCost::new(model, points_3d.clone(), points_2d.clone());
+
+        // Initial parameters
+        let param = vec![
+            self.model.intrinsics.fx,
+            self.model.intrinsics.fy,
+            self.model.intrinsics.cx,
+            self.model.intrinsics.cy,
+            self.model.alpha,
+            self.model.xi,
+        ];
+
+        let init_param: DVector<f64> = DVector::from_vec(param);
+
+        // Configure the Gauss-Newton solver
+        let solver = GaussNewton::new();
+
+        // Setup executor with the solver and cost function
+        let executor_builder = Executor::new(cost_function, solver)
+            .configure(|state| state.param(init_param).max_iters(100))
+            .add_observer(SlogLogger::term(), ObserverMode::NewBest);
+
+        if verbose {
+            info!("Starting optimization with Gauss-Newton...");
+        }
+
+        let res = executor_builder
+            .run()
+            .map_err(|e| CameraModelError::NumericalError(e.to_string()))?;
+
+        if verbose {
+            info!("Optimization finished: \n{}", res);
+            info!("Termination status: {:?}", res.state().termination_status);
+        }
+
+        let best_params_dvec = res.state().get_best_param().unwrap().clone();
+
+        // Update model parameters from the final parameters
+        self.model.intrinsics.fx = best_params_dvec[0];
+        self.model.intrinsics.fy = best_params_dvec[1];
+        self.model.intrinsics.cx = best_params_dvec[2];
+        self.model.intrinsics.cy = best_params_dvec[3];
+        self.model.alpha = best_params_dvec[4];
+        self.model.xi = best_params_dvec[5];
+
+        // Validate the optimized parameters
+        self.model.validate_params()?;
+
+        Ok(())
+    }
+
+    // linear_estimation will be fully implemented here when it's moved from CameraModel trait
+     fn linear_estimation(
+        intrinsics: &Intrinsics,
+        resolution: &Resolution,
+        points_2d: &Matrix2xX<f64>,
+        points_3d: &Matrix3xX<f64>,
+    ) -> Result<Self, CameraModelError>
+    where
+        Self: Sized {
+        // Check if the number of 2D and 3D points match
+        if points_2d.ncols() != points_3d.ncols() {
+            return Err(CameraModelError::InvalidParams(
+                "Number of 2D and 3D points must match".to_string(),
+            ));
+        }
+
+        // Initialize with xi = 0.0
+        let xi = 0.0;
+
+        // Set up the linear system to solve for alpha
+        let num_points = points_2d.ncols();
+        let mut a = nalgebra::DMatrix::zeros(num_points * 2, 1);
+        let mut b = nalgebra::DVector::zeros(num_points * 2);
+
+        for i in 0..num_points {
+            let x = points_3d[(0, i)];
+            let y = points_3d[(1, i)];
+            let z = points_3d[(2, i)];
+            let u = points_2d[(0, i)];
+            let v = points_2d[(1, i)];
+
+            let d = (x * x + y * y + z * z).sqrt();
+            let u_cx = u - intrinsics.cx;
+            let v_cy = v - intrinsics.cy;
+
+            a[(i * 2, 0)] = u_cx * (d - z);
+            a[(i * 2 + 1, 0)] = v_cy * (d - z);
+
+            b[i * 2] = (intrinsics.fx * x) - (u_cx * z);
+            b[i * 2 + 1] = (intrinsics.fy * y) - (v_cy * z);
+        }
+
+        // Solve the linear system using SVD
+        let svd = a.svd(true, true);
+        let alpha = match svd.solve(&b, 1e-10) {
+            Ok(sol) => sol[0], // Handle the successful case
+            Err(err_msg) => {
+                return Err(CameraModelError::NumericalError(err_msg.to_string()));
+            }
+        };
+
+        // Clamp alpha to valid range (0, 1]
+        let alpha = alpha;
+
+        let model = DoubleSphereModel {
+            intrinsics: intrinsics.clone(),
+            resolution: resolution.clone(),
+            alpha,
+            xi,
+        };
+
+        // Validate parameters
+        model.validate_params()?;
+
+        Ok(model)
+    }
 }
 
 /// Implementation of the Operator trait for Gauss-Newton optimization.
@@ -113,7 +250,7 @@ impl Jacobian for DoubleSphereOptimizationCost {
     fn jacobian(&self, p: &Self::Param) -> Result<Self::Jacobian, Error> {
         let num_points = self.points3d.ncols();
         let mut jacobian = DMatrix::zeros(num_points * 2, 6); // 2 residuals per point, 6 parameters
-        // Ensure DoubleSphereModel::new is public or accessible
+                                                              // Ensure DoubleSphereModel::new is public or accessible
         let model = DoubleSphereModel::new(&p)?;
         let mut counter = 0;
 
@@ -269,11 +406,11 @@ impl Hessian for DoubleSphereOptimizationCost {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::camera::{CameraModel, Intrinsics, Resolution, DoubleSphereModel as DSCameraModel}; // Renamed to avoid conflict
+    use crate::camera::{CameraModel, DoubleSphereModel as DSCameraModel, Intrinsics, Resolution}; // Renamed to avoid conflict
     use crate::optimization::Optimizer;
-    use nalgebra::{Matrix2xX, Matrix3xX, Vector3, DVector};
     use approx::assert_relative_eq;
-    use log::info; // For logging in tests if needed
+    use log::info;
+    use nalgebra::{DVector, Matrix2xX, Matrix3xX, Vector3}; // For logging in tests if needed
 
     // Helper to get a default model, similar to the one in samples/double_sphere.yaml
     // This is the camera model, not the Optimizer struct itself.
@@ -285,12 +422,15 @@ mod tests {
                 cx: 365.8121721753254,
                 cy: 249.3555778487899,
             },
-            resolution: Resolution { width: 752, height: 480 },
+            resolution: Resolution {
+                width: 752,
+                height: 480,
+            },
             alpha: 0.5657413673629862,
             xi: -0.24425190195168348,
         }
     }
-    
+
     // Placeholder for geometry::sample_points or a simplified version
     // For now, this will be very basic. A proper test would need more realistic points.
     fn sample_points_for_ds_model(
@@ -312,9 +452,11 @@ mod tests {
                 points_2d_vec.push(p2d);
             }
         }
-        Matrix2xX::from_columns(&points_2d_vec), Matrix3xX::from_columns(&points_3d_vec)
+        (
+            Matrix2xX::from_columns(&points_2d_vec),
+            Matrix3xX::from_columns(&points_3d_vec),
+        )
     }
-
 
     #[test]
     fn test_double_sphere_optimization_cost_basic() {
@@ -322,8 +464,11 @@ mod tests {
 
         // Generate sample points using the camera model
         let (points_2d, points_3d) = sample_points_for_ds_model(&model_camera, 5);
-        
-        assert!(points_3d.ncols() > 0, "Need at least one valid point for testing cost function.");
+
+        assert!(
+            points_3d.ncols() > 0,
+            "Need at least one valid point for testing cost function."
+        );
 
         // Construct cost struct
         let cost = DoubleSphereOptimizationCost::new(points_3d.clone(), points_2d.clone());
@@ -343,7 +488,10 @@ mod tests {
         assert_eq!(residuals.len(), 2 * points_3d.ncols());
         // For a perfect model and points, residuals should be near zero.
         // Here, points are projected by the same model, so they should be very small.
-        assert!(residuals.iter().all(|&v| v.abs() < 1e-6), "Residuals should be near zero for perfect model");
+        assert!(
+            residuals.iter().all(|&v| v.abs() < 1e-6),
+            "Residuals should be near zero for perfect model"
+        );
 
         // Test cost
         let c = cost.cost(&p).unwrap();
@@ -359,8 +507,10 @@ mod tests {
         let grad = cost.gradient(&p).unwrap();
         assert_eq!(grad.len(), 6);
         // For perfect model and points, gradient should be near zero.
-        assert!(grad.norm() < 1e-5, "Gradient norm should be near zero for perfect model");
-
+        assert!(
+            grad.norm() < 1e-5,
+            "Gradient norm should be near zero for perfect model"
+        );
 
         // Test hessian (J^T * J)
         let hess = cost.hessian(&p).unwrap();
@@ -373,9 +523,11 @@ mod tests {
     fn test_double_sphere_optimize_trait_method() {
         let reference_model = get_sample_camera_model();
         let (points_2d, points_3d) = sample_points_for_ds_model(&reference_model, 50);
-        
-        assert!(points_3d.ncols() > 10, "Need sufficient points for optimization test.");
 
+        assert!(
+            points_3d.ncols() > 10,
+            "Need sufficient points for optimization test."
+        );
 
         let mut noisy_model = DSCameraModel {
             intrinsics: Intrinsics {
@@ -386,29 +538,52 @@ mod tests {
             },
             resolution: reference_model.resolution.clone(),
             alpha: (reference_model.alpha * 0.90).max(0.1).min(0.99), // Increased noise
-            xi: reference_model.xi * 0.8, // Increased noise
+            xi: reference_model.xi * 0.8,                             // Increased noise
         };
 
         // Call optimize using the Optimizer trait
         let optimize_result = Optimizer::optimize(&mut noisy_model, &points_3d, &points_2d, false); // verbose = false
-        assert!(optimize_result.is_ok(), "Optimization failed: {:?}", optimize_result.err());
+        assert!(
+            optimize_result.is_ok(),
+            "Optimization failed: {:?}",
+            optimize_result.err()
+        );
 
         // Compare optimized parameters with reference_model
         // Allow for some tolerance as optimization might not be perfect
-        assert_relative_eq!(noisy_model.intrinsics.fx, reference_model.intrinsics.fx, epsilon = 10.0); // Looser epsilon
-        assert_relative_eq!(noisy_model.intrinsics.fy, reference_model.intrinsics.fy, epsilon = 10.0);
-        assert_relative_eq!(noisy_model.intrinsics.cx, reference_model.intrinsics.cx, epsilon = 10.0);
-        assert_relative_eq!(noisy_model.intrinsics.cy, reference_model.intrinsics.cy, epsilon = 10.0);
+        assert_relative_eq!(
+            noisy_model.intrinsics.fx,
+            reference_model.intrinsics.fx,
+            epsilon = 10.0
+        ); // Looser epsilon
+        assert_relative_eq!(
+            noisy_model.intrinsics.fy,
+            reference_model.intrinsics.fy,
+            epsilon = 10.0
+        );
+        assert_relative_eq!(
+            noisy_model.intrinsics.cx,
+            reference_model.intrinsics.cx,
+            epsilon = 10.0
+        );
+        assert_relative_eq!(
+            noisy_model.intrinsics.cy,
+            reference_model.intrinsics.cy,
+            epsilon = 10.0
+        );
         assert_relative_eq!(noisy_model.alpha, reference_model.alpha, epsilon = 0.1); // Looser
-        assert_relative_eq!(noisy_model.xi, reference_model.xi, epsilon = 0.1);       // Looser
+        assert_relative_eq!(noisy_model.xi, reference_model.xi, epsilon = 0.1); // Looser
     }
-    
+
     #[test]
     fn test_double_sphere_linear_estimation_optimizer_trait() {
         let reference_model = get_sample_camera_model();
         let (points_2d, points_3d) = sample_points_for_ds_model(&reference_model, 20);
 
-        assert!(points_3d.ncols() > 0, "Need points for linear estimation test.");
+        assert!(
+            points_3d.ncols() > 0,
+            "Need points for linear estimation test."
+        );
 
         let estimated_model_result = DSCameraModel::linear_estimation(
             &reference_model.intrinsics, // True intrinsics
@@ -417,7 +592,11 @@ mod tests {
             &points_3d,
         );
 
-        assert!(estimated_model_result.is_ok(), "Linear estimation failed: {:?}", estimated_model_result.err());
+        assert!(
+            estimated_model_result.is_ok(),
+            "Linear estimation failed: {:?}",
+            estimated_model_result.err()
+        );
         let estimated_model = estimated_model_result.unwrap();
 
         // Linear estimation for DoubleSphere typically estimates alpha, xi is often set to 0 or a fixed value.
@@ -425,11 +604,27 @@ mod tests {
         // So, we compare alpha and check if xi is close to 0.
         assert_relative_eq!(estimated_model.alpha, reference_model.alpha, epsilon = 0.2); // Linear estimation is an approximation
         assert_relative_eq!(estimated_model.xi, 0.0, epsilon = 1e-9); // Expect xi to be zero from linear_estimation impl
-        
+
         // Intrinsics should remain the same as input
-        assert_relative_eq!(estimated_model.intrinsics.fx, reference_model.intrinsics.fx, epsilon = 1e-9);
-        assert_relative_eq!(estimated_model.intrinsics.fy, reference_model.intrinsics.fy, epsilon = 1e-9);
-        assert_relative_eq!(estimated_model.intrinsics.cx, reference_model.intrinsics.cx, epsilon = 1e-9);
-        assert_relative_eq!(estimated_model.intrinsics.cy, reference_model.intrinsics.cy, epsilon = 1e-9);
+        assert_relative_eq!(
+            estimated_model.intrinsics.fx,
+            reference_model.intrinsics.fx,
+            epsilon = 1e-9
+        );
+        assert_relative_eq!(
+            estimated_model.intrinsics.fy,
+            reference_model.intrinsics.fy,
+            epsilon = 1e-9
+        );
+        assert_relative_eq!(
+            estimated_model.intrinsics.cx,
+            reference_model.intrinsics.cx,
+            epsilon = 1e-9
+        );
+        assert_relative_eq!(
+            estimated_model.intrinsics.cy,
+            reference_model.intrinsics.cy,
+            epsilon = 1e-9
+        );
     }
 }
