+++ src/optimization/mod.rs
@@ -4,9 +4,12 @@ pub mod double_sphere; // Added this line
 pub mod kannala_brandt; // Added this line
 pub mod rad_tan; // Added this line
 
-use nalgebra::{Matrix2xX, Matrix3xX};
+pub use double_sphere::DoubleSphereOptimizationCost;
+pub use kannala_brandt::KannalaBrandtOptimizationCost;
+pub use rad_tan::RadTanOptimizationCost;
+
 use crate::camera::{CameraModelError, Intrinsics, Resolution};
-use argmin::core::{CostFunction, Gradient, Hessian, Jacobian, Operator};
+use nalgebra::{Matrix2xX, Matrix3xX};
 
 pub trait Optimizer {
     fn optimize(
@@ -24,17 +27,4 @@ pub trait Optimizer {
     ) -> Result<Self, CameraModelError>
     where
         Self: Sized;
-}
-
-pub trait OptimizationCost:
-    CostFunction<Param = Self::Param, Output = Self::Output>
-    + Gradient<Param = Self::Param, Gradient = Self::Jacobian> // Or define a separate associated type for Gradient if it's not the Jacobian
-    + Hessian<Param = Self::Param, Hessian = Self::Hessian>
-    + Jacobian<Param = Self::Param, Jacobian = Self::Jacobian>
-    + Operator<Param = Self::Param, Output = Self::Output>
-{
-    type Param;
-    type Output;
-    type Jacobian; // As per argmin::core::Jacobian
-    type Hessian; // As per argmin::core::Hessian
 }
